feat: changed data list layout
