+++ src/Components/App/App.style.module.css
@@ -8,10 +8,14 @@
     flex-direction: column;
     gap: 40px;
     margin: 0 auto;
-    max-height: calc(100vh - 40px);
-    overflow-y: auto;
+    height: calc(100vh - 40px);
     padding-right: 17px;
     box-sizing: content-box;
+}
+
+.contentContainer {
+    flex-grow: 1;
+    height: calc(100vh - 35px - 40px - 40px);
 }
 
 .rightSide {
