+++ src/Components/Loader/Loader.styles.module.css
@@ -5,5 +5,7 @@
     justify-content: center;
     align-items: center;
     color: #0891b2;
-    background-color: rgba(255, 255, 255, 0.8);
+    font-size: 32px;
+    font-weight: bold;
+    background-color: rgba(255, 255, 255, 0.9);
 }
