+++ src/Components/Pagination/Pagination.styles.module.css
@@ -1,5 +1,4 @@
 .wrapper {
-    margin-top: 40px;
     width: 100%;
     display: flex;
     gap: 40px;
