+++ src/Components/SearchList/SearchList.styles.module.css
@@ -1,6 +1,13 @@
+.container {
+    height: 100%;
+    display: flex;
+    flex-direction: column;
+    gap: 20px;
+}
+
 .wrapper {
     position: relative;
-    min-height: 30px;
+    overflow-y: scroll;
     display: grid;
     grid-template-columns: repeat(auto-fill, minmax(200px, 1fr));
     gap: 20px;
