+++ src/Components/SearchList/SearchList.tsx
@@ -1,11 +1,9 @@
 import React, { useContext } from "react";
 import classes from "./SearchList.styles.module.css";
 import Loader from "../Loader/Loader";
-import { IMG_URL } from "../API/api.constants";
-import noImagePath from "../../assets/images/no-image.jpg";
 import Pagination from "../Pagination/Pagination";
-import { Link, useLocation } from "react-router-dom";
 import { AppContext } from "../AppContextProvider/AppContextProvider";
+import { SearchListItem } from "./SearchListItem";
 
 const SearchList: React.FC = () => {
   const {
@@ -14,44 +12,24 @@ const SearchList: React.FC = () => {
     pagination,
   } = useContext(AppContext);
 
-  const location = useLocation();
-
   return (
-    <div>
+    <div className={classes.container}>
       {searchList && searchList.length === 0 && (
-        <div>The search has not given any results</div>
+        <div>Поиск не дал результатов. Попробуйте еще раз.</div>
       )}
       <div className={classes.wrapper}>
         {searchList &&
           searchList.length > 0 &&
           searchList.map((item) => {
             const releaseData = new Date(item.release_date);
-
             return (
-              <Link
-                to={"/" + item.id.toString() + location.search}
+              <SearchListItem
+                id={item.id}
                 key={item.id}
-                className={classes.itemWrapper}
-              >
-                <div className={classes.posterWrapper}>
-                  <img
-                    src={
-                      item.poster_path
-                        ? `${IMG_URL}/${item.poster_path}`
-                        : noImagePath
-                    }
-                    alt={item.title}
-                  />
-                </div>
-                <div className={classes.textBlock}>
-                  <div>{item.title}</div>
-                  <div className={classes.itemCountry}>
-                    {releaseData.toLocaleDateString(undefined, {
-                      year: "numeric",
-                    })}
-                  </div>
-                </div>
-              </Link>
+                title={item.title}
+                poster_path={item.poster_path}
+                releaseData={releaseData}
+              />
             );
           })}
       </div>
