+++ src/Components/SearchList/SearchListItem.types.ts
@@ -1,0 +1,6 @@
+export interface typeSearchListItemProps {
+  id: number;
+  poster_path?: string;
+  title: string;
+  releaseData: Date;
+}
